requirements.txt
llama-index>=0.11.0
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma

chromadb==0.5.5
python-dotenv==1.0.1
pypdf==4.3.1

posthog<4
